刪除不必要檔案（本模組已停止維護）

+++ sql/mysql.sql
@@ -6,7 +6,7 @@ CREATE TABLE tad_cbox (
   `ip` varchar(255) NOT NULL default '' ,
   `only_root` enum('0','1') NOT NULL default '0' ,
   `root_msg` text NOT NULL ,
-  `uid` smallint(6) unsigned NOT NULL default 0 ,
+  `uid` mediumint(8) unsigned NOT NULL default 0 ,
   `box_sn` smallint(6) unsigned NOT NULL default 0 ,
   `re_sn` mediumint(9) unsigned NOT NULL default 0 ,
   PRIMARY KEY  (`sn`),
